CAPrimaryVertexContextInitializer .h/.cxx creation
